[1129_장민] fix_학점_실수형

+++ main/models.py
@@ -42,7 +42,7 @@ class UserLecture(models.Model): # 학생성적정보파일 업로드 시 저장
     major = models.CharField(max_length=50)
     year = models.IntegerField()
     semester = models.CharField(max_length=20)
-    grade = models.IntegerField()
+    grade = models.FloatField()
 
     class Meta:
        managed = False
